Move bundle underneath controllers
